% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarLists4CMIP5.R
\name{GetVarLists4CMIP5}
\alias{GetVarLists4CMIP5}
\title{GetVarLists4CMIP5}
\usage{
GetVarLists4CMIP5(basedir, coln)
}
\arguments{
\item{basedir}{Directory path containing NetCDF files}

\item{coln}{Column number for extracting available list (1: variables, 2: time-step, 3: models, 4: scenarios, 5: ensambles, 6: periods)}
}
\value{
varnm List of available variable
}
\description{
Extract necessary list from NetCDF file names according to column number (coln: 1: variables, 2: time-step, 3: models, 4: scenarios, 5: ensambles, 6: periods)
}
