% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWorkingDir.R
\name{SetWorkingDir}
\alias{SetWorkingDir}
\title{SetWorkingDir}
\usage{
SetWorkingDir(wdir)
}
\arguments{
\item{wdir}{Working directory path}
}
\description{
Create a working directory if it is not available
}
